Move animation types to separate file
