+++ desktop/packages/mullvad-vpn/src/renderer/lib/components/animate/types.ts
@@ -1,0 +1,10 @@
+export type Animation = FadeAnimation | WipeAnimation;
+
+export type FadeAnimation = {
+  type: 'fade';
+};
+
+export type WipeAnimation = {
+  type: 'wipe';
+  direction: 'vertical';
+};
